Rename package duck.util to duck.common
Utils class, containing common utility functions used across various classes was packaged with util package.

Let's:
- Rename the package to duck.common
- Create a Message class to store all the common messages used across the application

This allows the duck.common package to contain all common classes and functions used across the application.
