+++ src/main/java/duck/commands/UnmarkCommand.java
@@ -4,7 +4,7 @@ import duck.data.TaskList;
 import duck.data.exception.DuckException;
 import duck.storage.Storage;
 import duck.ui.Ui;
-import duck.util.Utils;
+import duck.common.Utils;
 
 /**
  * Represents a command to unmark a task as incomplete.
