+++ src/main/java/duck/common/Message.java
@@ -1,0 +1,10 @@
+package duck.common;
+
+/**
+ * Contains messages to be used across multiple classes.
+ */
+public class Message {
+
+    public static final String INVALID_INDEX_FORMAT = "Oops! you have to indicate a valid task index!\n";
+    public static final String INDEX_OUT_OF_BOUNDS = "Oops! Index out of bound :( Input a valid task index.\n";
+}
